+++ GameOfLife.Data/Entities/Game/GameHandler.cs
@@ -1,6 +1,7 @@
 ﻿using GameOfLife.Data.Interfaces;
+using GameOfLife.Data.Interfaces.Game;
 
-namespace GameOfLife.Data.Entities
+namespace GameOfLife.Data.Entities.Game
 {
     public class GameHandler(IRule rule) : IGameHandler
     {
@@ -18,9 +19,9 @@ namespace GameOfLife.Data.Entities
 
             var nextMap = new Map(map);
 
-            Parallel.For(0, height, i=>
+            Parallel.For(0, height, i =>
             {
-                for(int j = 0; j < length; j++)
+                for (int j = 0; j < length; j++)
                 {
                     var cell = map.GetCell(i, j);
                     var aliveNeighbours = CountAliveNeighbours(i, j, map);
@@ -29,7 +30,7 @@ namespace GameOfLife.Data.Entities
 
                     nextMap.SetCell(new Cell(i, j, nextState));
                 }
-            });      
+            });
 
             return nextMap;
         }
@@ -45,9 +46,9 @@ namespace GameOfLife.Data.Entities
         {
             int count = 0;
 
-            for(int i = x - 1; i <= x + 1; i++)
+            for (int i = x - 1; i <= x + 1; i++)
             {
-                for(int j = y - 1; j <= y + 1; j++)
+                for (int j = y - 1; j <= y + 1; j++)
                 {
                     if (i == x && j == y || i < 0 || j < 0 || i >= map.Height || j >= map.Length)
                         continue;
