+++ GameOfLife.Data/Entities/Menus/Menu.cs
@@ -1,7 +1,7 @@
 ﻿using GameOfLife.Data.Interfaces;
 
 
-namespace GameOfLife.Data.Entities
+namespace GameOfLife.Data.Entities.Menus
 {
     public abstract class Menu
     {
