+++ GameOfLife.Data/Entities/Menus/MenuItem.cs
@@ -1,6 +1,6 @@
 ﻿using GameOfLife.Data.Interfaces;
 
-namespace GameOfLife.Data.Entities
+namespace GameOfLife.Data.Entities.Menus
 {
     public class MenuItem : IMenuItem
     {
