+++ GameOfLife.Data/Interfaces/Game/IGameService.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace GameOfLife.Data.Interfaces
+namespace GameOfLife.Data.Interfaces.Game
 {
     public interface IGameService
     {
