+++ GameOfLife.Data/Interfaces/UI/IPrintable.cs
@@ -1,4 +1,4 @@
-﻿namespace GameOfLife.Data.Interfaces
+﻿namespace GameOfLife.Data.Interfaces.UI
 {
     public interface IPrintable
     {
